Split out transparent water for US
