+++ src/st/np3/e_transparent_water.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "np3.h"
+
+#include "../e_transparent_water.h"
